miniconda on home/download

+++ install.sh
@@ -14,8 +14,20 @@ else
     url="https://repo.continuum.io/miniconda/Miniconda3-latest-$infix-x86_64.sh"
 fi
 
-wget $url -O miniconda.sh;
-bash miniconda.sh -b -p $HOME/miniconda
+pushd .
+cd
+mkdir -p download
+cd download
+echo "Cached in $HOME/download :"
+ls -l
+echo
+if [[ ! -f miniconda.sh ]]
+then
+    wget $url -O miniconda.sh
+    chmod +x miniconda.sh
+    bash miniconda.sh -b -p $HOME/miniconda
+fi
+popd
 
 export PATH="$HOME/miniconda/bin:$PATH"
 hash -r
